# ZKVM-1186: Use wide types in poseidon2