--- a/risc0/circuit/rv32im-v2/Cargo.toml
+++ b/risc0/circuit/rv32im-v2/Cargo.toml
@@ -52,6 +52,7 @@ smallvec = { version = "1.13", optional = true }
 [dev-dependencies]
 clap = { version = "4.5", features = ["derive"] }
 flate2 = "1.0"
+rstest = "0.24.0"
 tar = "0.4"
 test-log = { version = "0.2", default-features = false, features = ["trace"] }
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }