This improves my poseidon2 benchmark from 1.77s to 1.43s. 

With this code if we turn on the vectorized instructions it goes down from 1.43s to 1.22s. We could do some runtime detection of the processor features in a future change maybe.