@bobbobbio 2025-03-06
these use of wide types don't help the other poseidon2 implementation